applyLinks separated
