+++ src/ui/main.ts
@@ -4,7 +4,8 @@ import url from 'url';
 import path from 'path';
 import electron from 'electron';
 
-import * as dropTheRopeRunner from '../dropTheRopeRunner';
+import { applyLinks } from '../utils/apply-links';
+import * as yarle from '../yarle';
 import { loggerInfo } from '../utils/loggerInfo';
 
 import { store } from './store';
@@ -177,8 +178,9 @@ electron.ipcMain.on('configurationUpdated', (event: any, data: any) => {
 
 electron.ipcMain.on('startConversion', async (event: any, data: any) => {
   const settings = mapSettingsToYarleOptions();
-  // await yarle.dropTheRope(settings);
-  await dropTheRopeRunner.run(settings);
+  const outputNotebookFolders = await yarle.dropTheRope(settings);
+  // apply internal links
+  applyLinks(settings, outputNotebookFolders);
 
 });
